// src/public/dto/post/view-post.dto.ts
export class ViewPostDto {
  id: string;
  title: string;
  shortDescription: string;
  content: string;
  blogId: string;
  blogName: string;
  createdAt: string;
  extendedLikesInfo: {
    likesCount: number;
    dislikesCount: number;
    myStatus: string;
    newestLikes: likeInfo[];
  };
}

export class likeInfo {
  addedAt: string;
  userId: string;
  login: string;
}